feat: Adds shrine-ftp gem support
* new sample shrine_ftp.rb initializer for if you need to upload via ftp
* bugfix deleting a file in local directory when it doesn't exist

+++ app/services/flow_service.rb
@@ -56,7 +56,9 @@ class FlowService
   def clean
     FileUtils.rm_rf chunk_file_directory
     [final_file_path, final_flac_path, mp3_file_path, he_aac_file_path].each do |file|
-      FileUtils.remove file
+      if File.exist?(file)
+        FileUtils.remove file
+      end
     end
     true
   end
